+++ grid3.h
@@ -20,7 +20,7 @@ class Grid3 {
   // 8 grid points around projection of each point:
   // first: array of index in _s of the points
   // second: array of weights
-  std::pair<int[8], double[8]>* _projection;
+  std::pair<int[8], float[8]>* _projection;
   
   // 19 grid points around each point and itself
   Neighbours* _neighbour;
@@ -28,29 +28,29 @@ class Grid3 {
   // density
   Function _rho;
 
-  double _t;
-  double _h;
-  double _delta;
+  float _t;
+  float _h;
+  float _delta;
 
   public:
   Grid3(Interval x = {0,10},
     Interval y = {0,10},
     Interval z = {0,10},
-    double T = 40,
-    double h = 0.1);
-  Grid3(std::string surfFile, std::string realFile, double T);
+    float T = 40.,
+    float h = 0.1);
+  Grid3(std::string surfFile, std::string realFile, float T);
   ~Grid3();
 
   void saveRelations(std::string filename = "simulation");
   void saveSurface(std::string filename = "simulation");
   void saveRho(std::string filename = "simulation");
 
-  void evolve(double dt = 0);
+  void evolve(float dt = 0);
   void project();
 
   private:
   //signed distance from the sphere
-  double phi(double x, double y, double z);
+  float phi(float x, float y, float z);
 
   //create the first rho value
   void createRho();
@@ -71,13 +71,13 @@ class Grid3 {
   bool loadRho(std::string filename = "simulation");
 
   //first derivative along one direction
-  double der1(int pointIndex, int direction);
+  float der1(int pointIndex, int direction);
 
   //second derivative along the same direction
-  double der2(int pointIndex, int direction);
+  float der2(int pointIndex, int direction);
 
   //second mix derivative along two different directions
-  double derij(int pointIndex, int dir1, int dir2);
+  float derij(int pointIndex, int dir1, int dir2);
 
   //logic end of int between -1 and +1
   int sgn(int val1, int val2);
